Remove between-threads acquire and release, return null on empty q

+++ MaximNigmatulin/Task2/src/main/kotlin/mechanism/DataStore.kt
@@ -2,43 +2,39 @@ package pc.mechanism
 
 import java.util.Queue
 import java.util.LinkedList
-import java.util.concurrent.Semaphore
+import java.util.concurrent.locks.ReentrantLock
+import kotlin.concurrent.withLock
+
 
 class DataStore<T>(val log: (String) -> Unit = {}) {
     private val queue: Queue<T> = LinkedList()
-    private val lock = Semaphore(1)
-    private val emptyQueueLock = Semaphore(1)
-
-    init {
-        emptyQueueLock.acquire()
-    }
+    private val lock = ReentrantLock()
 
     fun push(obj: T) {
-        log("waiting to get permit to PUSH")
-        lock.acquire()
-        log("locked to push $obj")
+        log("trying to push $obj")
+        while (!lock.tryLock()) {
+        }
+
         queue.add(obj)
         val snapshot = expose()
         log("added $obj -> $snapshot")
-        log("unlock")
-
-        lock.release()
-        emptyQueueLock.release()
+        lock.unlock()
     }
 
-    fun pop(): T {
-        log("waiting to get permit to POP")
-        emptyQueueLock.acquire()
-        lock.acquire()
-
-        log("lock to remove last")
+    fun pop(): T? {
+        log("trying to pop")
+        while (!lock.tryLock()) {
+        }
+        if (queue.isEmpty()) {
+            lock.unlock()
+            return null
+        }
         val toReturn = queue.remove()
         val snapshot = expose()
 
         log("rm $toReturn -> $snapshot")
-        log("unlock")
 
-        lock.release()
+        lock.unlock()
         return toReturn
     }
